refactor(scripts): separate /storybook and /webpack and introduce barell api within /webpack
