+++ apps/theming-designer/webpack.serve.config.js
@@ -1,5 +1,4 @@
-const resources = require('../../scripts/webpack/webpack-resources');
-const getResolveAlias = require('../../scripts/webpack/getResolveAlias');
+const { getResolveAlias, resources } = require('../../scripts/webpack');
 
 module.exports = resources.createServeConfig(
   {
