+++ scripts/storybook/tsconfig.json
@@ -3,6 +3,5 @@
   "compilerOptions": {
     "types": ["node", "jest"]
   },
-  "include": ["*"],
-  "files": ["../../typings/ignore-not-found-export-webpack-plugin/index.d.ts"]
+  "include": ["*"]
 }
