+++ scripts/webpack/README.md
@@ -1,0 +1,9 @@
+# webpack
+
+Webpack utils and helpers for v8 (`@fluentui/react`) packages
+
+## Usage
+
+```js
+import { resources, getResolveAlias } from '@fluentui/scripts/webpack';
+```
