+++ scripts/webpack/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  ...require('./storybook-webpack.config'),
+  ...require('./getResolveAlias'),
+  resources: require('./webpack-resources'),
+};
